[Step Function] 6.1 Move common files to common/ folder
